Moved ZookeeperConnection
