+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/TaskManager.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.engine.tasks;
 
-import ai.grakn.engine.tasks.manager.ZookeeperConnection;
+import ai.grakn.engine.tasks.connection.ZookeeperConnection;
 import ai.grakn.engine.tasks.storage.TaskStateGraphStore;
 import ai.grakn.engine.tasks.storage.TaskStateZookeeperStore;
 import ai.grakn.engine.GraknEngineConfig;
