+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/connection/ZookeeperConnection.java
@@ -14,10 +14,9 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
- *
  */
 
-package ai.grakn.engine.tasks.manager;
+package ai.grakn.engine.tasks.connection;
 
 import ai.grakn.engine.GraknEngineConfig;
 import org.apache.curator.framework.CuratorFramework;
