+++ grakn-test/src/test/java/ai/grakn/test/engine/lock/LockProviderTest.java
@@ -21,7 +21,7 @@ package ai.grakn.test.engine.lock;
 import ai.grakn.engine.lock.LockProvider;
 import ai.grakn.engine.lock.NonReentrantLock;
 import ai.grakn.engine.lock.ZookeeperLock;
-import ai.grakn.engine.tasks.manager.ZookeeperConnection;
+import ai.grakn.engine.tasks.connection.ZookeeperConnection;
 import java.util.concurrent.locks.Lock;
 import org.apache.curator.framework.CuratorFramework;
 import org.junit.After;
