+++ grakn-test/src/test/java/ai/grakn/test/engine/lock/LockTest.java
@@ -20,7 +20,7 @@ package ai.grakn.test.engine.lock;
 
 import ai.grakn.engine.lock.NonReentrantLock;
 import ai.grakn.engine.lock.ZookeeperLock;
-import ai.grakn.engine.tasks.manager.ZookeeperConnection;
+import ai.grakn.engine.tasks.connection.ZookeeperConnection;
 import ai.grakn.test.EngineContext;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
